requirements.txt
requests>=2.28.0
pandas>=1.5.0
PyYAML>=6.0
Jinja2>=3.1.0
pytest>=7.2.0
